% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense.R
\name{condense}
\alias{condense}
\title{condense}
\usage{
condense(pca, n, k)
}
\arguments{
\item{pca}{\code{prcomp} output from stats::prcomp (PCA)}

\item{n}{\code{integer} number of PCs to extract}

\item{k}{\code{integer} number of features per PC}
}
\value{
list
}
\description{
find n top PCs and k highest loading features for those PCs
}
