% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_different_pc.R
\name{most_different_pc}
\alias{most_different_pc}
\title{most_different_pc}
\usage{
most_different_pc(pca, label_vec, n)
}
\arguments{
\item{pca}{\code{prcomp} output from stats::prcomp}

\item{n}{\code{integer} number of PCs to extract.}
}
\value{
\code{list}
}
\description{
returns the sorted indices of first n PCs with the highest median differences across groups.
}
