% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{read_data}
\usage{
read_data(train, labels, features)
}
\arguments{
\item{train}{\code{character} path to training dataset (matrix) - must be .csv}

\item{labels}{\code{character} path to training dataset (labels) - must be .txt}

\item{features}{\code{character} path to training dataset (features) - must be .txt}
}
\value{
df \code{data.frame} returns a data.frame containing training data with labels and features.
}
\description{
read_data
}
\examples{
\dontrun {
dir("out/data_import") #show files in a directory
train <- "out/data_import/contract_binarytrain_data.csv"
labels <- "out/data_import/contract_binarytrain_labels.txt"
features <- "out/data_import/contract_binarytrain_vocabulary.txt"
df <- read_data(train, labels, features)
}
}
